add osx kraken, refactor oodle loaders
