+++ WolvenKit.Common/DDS/DDS_Metadata.cs
