+++ WolvenKit.Common/DDS/DDS_PIXELFORMAT.cs
