+++ WolvenKit.Core/Compression/KrakenNative.cs
@@ -2,7 +2,7 @@ using System.Runtime.InteropServices;
 
 namespace WolvenKit.Core.Compression;
 
-public static class KrakenNative
+internal static class KrakenNative
 {
     public static int Decompress(byte[] buffer, byte[] outputBuffer)
         => Kraken_Decompress(buffer, buffer.Length, outputBuffer, outputBuffer.Length);
@@ -10,33 +10,17 @@ public static class KrakenNative
     public static int Compress(byte[] buffer, byte[] outputBuffer, int level)
         => Kraken_Compress(buffer, buffer.Length, outputBuffer, level);
 
-    // #region osx
-    //
-    // [DllImport("lib/kraken.dylib", CallingConvention = CallingConvention.StdCall)]
-    // public static extern int Compress(
-    //     int algorithm,
-    //     IntPtr inputBuffer,
-    //     IntPtr outputBuffer,
-    //     int inputSize,
-    //     IntPtr src_window_base,
-    //     IntPtr lrm_org);
-    //
-    // [DllImport("lib/kraken.dylib", CallingConvention = CallingConvention.StdCall)]
-    // public static extern int GetCompressedBufferSizeNeeded(int size);
-    //
-    // [DllImport("lib/kraken.dylib", CallingConvention = CallingConvention.StdCall)]
-    // public static extern int Kraken_Decompress(
-    //     byte[] buffer,
-    //     long bufferSize,
-    //     byte[] outputBuffer,
-    //     long outputBufferSize);
-    //
-    // #endregion osx
 
-    #region Win
+#if _WINDOWS
+    private const string dllName = "lib/kraken.dll";
+#elif _OSX
+    private const string dllName = "lib/libkraken.dylib";
+#elif _LINUX
+    private const string dllName = "lib/libkraken.so";
+#endif
 
     // EXPORT int Kraken_Decompress(const byte* src, size_t src_len, byte* dst, size_t dst_len);
-    [DllImport("lib/kraken.dll", CallingConvention = CallingConvention.StdCall)]
+    [DllImport(dllName, CallingConvention = CallingConvention.StdCall)]
     private static extern int Kraken_Decompress(
         byte[] buffer,
         long bufferSize,
@@ -44,12 +28,10 @@ public static class KrakenNative
         long outputBufferSize);
 
     // EXPORT int Kraken_Compress(uint8* src, size_t src_len, byte* dst, int level);
-    [DllImport("lib/kraken.dll", CallingConvention = CallingConvention.StdCall)]
+    [DllImport(dllName, CallingConvention = CallingConvention.StdCall)]
     private static extern int Kraken_Compress(
        byte[] buffer,
        long bufferSize,
        byte[] outputBuffer,
        int level);
-
-    #endregion
 }
