+++ WolvenKit.MSTests/OozTests.cs
@@ -16,47 +16,11 @@ namespace WolvenKit.MSTests
         [ClassInitialize]
         public static void SetupClass(TestContext _)
         {
-            #region oodle
-            var s_config = new ConfigurationBuilder()
-                .AddJsonFile("appsettings.json")
-                .Build();
-
-            var s_gameDirectoryPath = "";
-            var cp77Dir = Environment.GetEnvironmentVariable("CP77_DIR", EnvironmentVariableTarget.User);
-            if (!string.IsNullOrEmpty(cp77Dir) && new DirectoryInfo(cp77Dir).Exists)
-            {
-                s_gameDirectoryPath = cp77Dir;
-            }
-            else
-            {
-                s_gameDirectoryPath = s_config.GetSection(s_gameDirectorySetting).Value;
-            }
-            var gameDirectory = new DirectoryInfo(s_gameDirectoryPath);
-            var gameBinDir = new DirectoryInfo(Path.Combine(gameDirectory.FullName, "bin", "x64"));
-            var oodleInfo = new FileInfo(Path.Combine(gameBinDir.FullName, "oo2ext_7_win64.dll"));
-            if (!oodleInfo.Exists)
-            {
-                Assert.Fail("Could not find oo2ext_7_win64.dll.");
-            }
             var ass = AppDomain.CurrentDomain.BaseDirectory;
-            var appOodleFileName = Path.Combine(ass, "oo2ext_7_win64.dll");
-            if (!File.Exists(appOodleFileName))
-            {
-                oodleInfo.CopyTo(appOodleFileName);
-            }
-
-            if (!OodleLib.Load(appOodleFileName))
+            if (!Oodle.Load())
             {
                 Assert.Fail("Could not load oo2ext_7_win64.dll.");
             }
-
-            var kraken = Path.Combine(ass,"lib", "kraken.dll");
-            if (!KrakenLib.Load(kraken))
-            {
-                Assert.Fail("Could not load kraken.dll.");
-            }
-
-            #endregion
         }
 
         [TestMethod]
@@ -85,7 +49,7 @@ namespace WolvenKit.MSTests
                 var outpath = Path.Combine(outdir, "oodle.txt");
                 File.WriteAllBytes(outpath, outBuffer1);
             }
-
+#if _WINDOWS
             // kraken
             var outBuffer2 = new byte[size];
             {
@@ -94,6 +58,7 @@ namespace WolvenKit.MSTests
                 File.WriteAllBytes(outpath2, outBuffer2);
             }
             Assert.IsTrue(Enumerable.SequenceEqual(outBuffer1, outBuffer2));
+#endif
         }
 
         [TestMethod]
@@ -111,20 +76,15 @@ namespace WolvenKit.MSTests
             var outdir = Path.Combine(Environment.CurrentDirectory, "ooz");
             Directory.CreateDirectory(outdir);
 
+
             // oodle
             IEnumerable<byte> outBuffer = new List<byte>();
-            var r = Oodle.Compress(inbuffer,ref outBuffer, false, (OodleLZNative.CompressionLevel)level);
+            var r = Oodle.Compress(inbuffer,ref outBuffer, false, (Oodle.CompressionLevel)level);
             var outpath = Path.Combine(outdir, $"{(int)level}_oodle.kark");
             var final = outBuffer.ToArray();
             File.WriteAllBytes(outpath, final);
 
-            //// oodle native
-            //var outBuffer1 = new byte[compressedBufferSizeNeeded];
-            //var r1 = OodleLZNative.Compress(OodleLZNative.Compressor.Kraken, inbuffer, inbuffer.Length, outBuffer1, (OodleLZNative.CompressionLevel)level);
-            //var outpath1 = Path.Combine(outdir, $"{(int)level}_noodle.kark");
-            //var final1 = outBuffer.ToArray();
-            //File.WriteAllBytes(outpath1, final1);
-
+#if _WINDOWS
             // kraken
             var outBuffer2 = new byte[compressedBufferSizeNeeded];
             //var outBuffer2 = new byte[inbuffer.Length + 65536];
@@ -132,10 +92,11 @@ namespace WolvenKit.MSTests
             var outpath2 = Path.Combine(outdir, $"{(int)level}_kraken.kark");
             var final2 = outBuffer2.Take(r2).ToArray();
             File.WriteAllBytes(outpath2, final2);
-            
-            //Assert.IsTrue(Enumerable.SequenceEqual(final, final1));
-            //Assert.IsTrue(Enumerable.SequenceEqual(final1, final2));
+
             Assert.IsTrue(Enumerable.SequenceEqual(final, final2));
+#endif
+
+
         }
     }
 }
